# swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "BuntesBegegnen.Api",
    "version": "1.0"
  },
  "paths": {
...
    "/activities": {
      "post": {
        "tags": [
          "Activities"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ActivityCreatedDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActivityCreatedDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActivityCreatedDto"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Activities"
        ],
        "parameters": [
          {
            "name": "minDate",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "maxDate",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "searchTerm",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "onlyRegistered",
            "in": "query",
            "schema": {
              "type": "boolean",
              "default": false
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ActivityDto"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ActivityDto"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ActivityDto"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/activities/{id}": {
      "get": {
        "tags": [
          "Activities"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ActivityDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActivityDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActivityDto"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Activities"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateActivityDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateActivityDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateActivityDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "delete": {
        "tags": [
          "Activities"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/activities/{id}/register": {
      "post": {
        "tags": [
          "Activities"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/activities/{id}/unregister": {
      "post": {
        "tags": [
          "Activities"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ActivityCategory": {
        "enum": [
          "Excursion",
          "Creativity",
          "Exercise"
        ],
        "type": "string"
      },
      "ActivityCreatedDto": {
        "required": [
          "id"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
      "ActivityDto": {
        "required": [
          "createdAt",
          "createdById",
          "id"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "createdById": {
            "type": "string"
          },
          "promoterId": {
            "type": "integer",
            "format": "int32"
          },
          "title": {
            "type": "string"
          },
          "visibility": {
            "$ref": "#/components/schemas/ActivityVisibility"
          },
          "category": {
            "$ref": "#/components/schemas/ActivityCategory"
          },
          "startTime": {
            "type": "string",
            "format": "date-time"
          },
          "endTime": {
            "type": "string",
            "format": "date-time"
          },
          "isAllDay": {
            "type": "boolean"
          },
          "maxNumberOfParticipants": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "registrationLocked": {
            "type": "boolean"
          },
          "location": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "recurrenceFrequency": {
            "$ref": "#/components/schemas/ActivityRecurrenceFrequency"
          },
          "recurrenceInterval": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "repeatUntil": {
            "type": "string",
            "format": "date",
            "nullable": true
          },
          "repeatCount": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "isInstance": {
            "type": "boolean"
          },
          "recurrenceDates": {
            "uniqueItems": true,
            "type": "array",
            "items": {
              "type": "string",
              "format": "date-time"
            }
          },
          "recurrenceByDay": {
            "uniqueItems": true,
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ActivityRecurrenceByDayDto"
            }
          },
          "recurrenceByMonthDay": {
            "uniqueItems": true,
            "type": "array",
            "items": {
              "type": "integer",
              "format": "int32"
            }
          },
          "recurrenceExceptions": {
            "uniqueItems": true,
            "type": "array",
            "items": {
              "type": "string",
              "format": "date-time"
            }
          },
          "isRegistered": {
            "type": "boolean"
          },
          "currentNumberOfParticipants": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
...
      "ActivityRecurrenceByDayDto": {
        "type": "object",
        "properties": {
          "ordinal": {
            "type": "integer",
            "format": "int32"
          },
          "dayOfWeek": {
            "$ref": "#/components/schemas/DayOfWeek"
          }
        },
        "additionalProperties": false
      },
...
      "ActivityRecurrenceFrequency": {
        "enum": [
          "None",
          "Weekly",
          "Monthly",
          "FixedDates"
        ],
        "type": "string"
      },
      "ActivityVisibility": {
        "enum": [
          "PrivateDraft",
          "SharedDraft",
          "Public"
        ],
        "type": "string"
      },
...
      "DayOfWeek": {
        "enum": [
          "Sunday",
          "Monday",
          "Tuesday",
          "Wednesday",
          "Thursday",
          "Friday",
          "Saturday"
        ],
        "type": "string"
      },
...
      "UpdateActivityDto": {
        "type": "object",
        "properties": {
          "promoterId": {
            "type": "integer",
            "format": "int32"
          },
          "title": {
            "maxLength": 100,
            "minLength": 0,
            "type": "string"
          },
          "visibility": {
            "$ref": "#/components/schemas/ActivityVisibility"
          },
          "category": {
            "$ref": "#/components/schemas/ActivityCategory"
          },
          "startTime": {
            "type": "string",
            "format": "date-time"
          },
          "endTime": {
            "type": "string",
            "format": "date-time"
          },
          "isAllDay": {
            "type": "boolean"
          },
          "maxNumberOfParticipants": {
            "maximum": 10000,
            "minimum": 0,
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "registrationLocked": {
            "type": "boolean"
          },
          "location": {
            "maxLength": 200,
            "minLength": 3,
            "type": "string"
          },
          "description": {
            "maxLength": 10000,
            "minLength": 10,
            "type": "string"
          },
          "recurrenceFrequency": {
            "$ref": "#/components/schemas/ActivityRecurrenceFrequency"
          },
          "recurrenceInterval": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "repeatUntil": {
            "type": "string",
            "format": "date",
            "nullable": true
          },
          "repeatCount": {
            "maximum": 1000,
            "minimum": 1,
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "recurrenceDates": {
            "uniqueItems": true,
            "type": "array",
            "items": {
              "type": "string",
              "format": "date-time"
            }
          },
          "recurrenceByDay": {
            "uniqueItems": true,
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ActivityRecurrenceByDayDto"
            }
          },
          "recurrenceByMonthDay": {
            "uniqueItems": true,
            "type": "array",
            "items": {
              "type": "integer",
              "format": "int32"
            }
          },
          "recurrenceExceptions": {
            "uniqueItems": true,
            "type": "array",
            "items": {
              "type": "string",
              "format": "date-time"
            }
          }
        },
        "additionalProperties": false
      },
...
    },
    "securitySchemes": {
      "Bearer": {
        "type": "http",
        "description": "Enter the authentication token (without the 'Bearer' prefix)",
        "scheme": "Bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "Bearer": [ ]
    }
  ]
}